Move /directions to /internal/directions
